+++ packages/terra-abstract-modal/src/AbstractModal.jsx
@@ -61,10 +61,6 @@ const propTypes = {
    */
   zIndex: PropTypes.oneOf(zIndexes),
   /**
-   * If set to true, the focus will be trapped within the modal content when the modal is open.
-   */
-  trapFocus: PropTypes.bool,
-  /**
    * @private
    * Callback function to set the reference of the element that will receive focus when the Slide content is visible.
    */
@@ -80,7 +76,6 @@ const defaultProps = {
   role: 'dialog',
   rootSelector: '#root',
   zIndex: '6000',
-  trapFocus: false,
 };
 
 const AbstractModal = (props) => {
@@ -97,7 +92,6 @@ const AbstractModal = (props) => {
     rootSelector,
     onRequestClose,
     zIndex,
-    trapFocus,
     ...customProps
   } = props;
 
@@ -160,7 +154,6 @@ const AbstractModal = (props) => {
         aria-modal="true"
         ref={modalElementRef}
         onKeyDown={handleKeydown}
-        trapFocus={trapFocus}
       >
         {children}
       </ModalContent>