+++ packages/terra-abstract-modal/src/_ModalContent.jsx
@@ -60,10 +60,6 @@ const propTypes = {
    */
   zIndex: PropTypes.oneOf(zIndexes),
   /**
-   * If set to true, the focus will be trapped within the modal content when the modal is open.
-   */
-  trapFocus: PropTypes.bool,
-  /**
    * @private
    * Callback function to set the reference of the element that will receive focus when the Slide content is visible.
    */
@@ -78,19 +74,8 @@ const defaultProps = {
   isScrollable: false,
   role: 'dialog',
   rootSelector: '#root',
-  trapFocus: false,
   zIndex: '6000',
 };
-
-/**
- * Use react-focus-lock library due to inability of focus-trap-react to handle children
- * within a React Portal: https://github.com/focus-trap/focus-trap-react/issues/27.
- */
-// eslint-disable-next-line react/prop-types
-const FocusTrapWrapper = ({ trapFocus, children }) => (trapFocus
-  ? <FocusLock as="div" className={cx('modal-content-focus-trap-container')}>{children}</FocusLock>
-  : children
-);
 
 const ModalContent = forwardRef((props, ref) => {
   const {
@@ -106,7 +91,6 @@ const ModalContent = forwardRef((props, ref) => {
     rootSelector,
     zIndex,
     setModalFocusElementRef,
-    trapFocus,
     ...customProps
   } = props;
 
@@ -161,7 +145,13 @@ const ModalContent = forwardRef((props, ref) => {
         role={role}
         ref={ref}
       >
-        <FocusTrapWrapper trapFocus={trapFocus}>
+        {
+          /**
+           * Use react-focus-lock library due to inability of focus-trap-react to handle children
+           * within a React Portal: https://github.com/focus-trap/focus-trap-react/issues/27.
+           */
+        }
+        <FocusLock as="div" className={cx('modal-content-focus-trap-container')}>
           <div className={modalContainerClassName} ref={setModalFocusElementRef} data-terra-abstract-modal-begin tabIndex="-1">
             <FormattedMessage id="Terra.AbstractModal.BeginModalDialog">
               {text => {
@@ -189,7 +179,7 @@ const ModalContent = forwardRef((props, ref) => {
               }}
             </FormattedMessage>
           </div>
-        </FocusTrapWrapper>
+        </FocusLock>
       </div>
     </React.Fragment>
   );